Reorganise block header
"The header of a PHP file may consist of a number of different blocks.
If present, each of the blocks below MUST be separated by a single blank
line, and MUST NOT contain a blank line. Each block MUST be in the order
listed below, although blocks that are not relevant may be omitted."